pyproject.toml
[project]
name = "smartsaber"
version = "0.1.0"
description = "Import Spotify playlists into Beat Saber with auto-generated maps"
requires-python = ">=3.10"

dependencies = [
    "spotipy>=2.23.0",
    "httpx>=0.27.0",
    "yt-dlp>=2024.0.0",
    "rapidfuzz>=3.0.0",
    "rich>=13.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
    "Pillow>=9.0.0",
    "questionary>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
smartsaber = "smartsaber.cli:main"

[tool.setuptools.packages.find]
include = ["smartsaber*"]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
